BuyerCounterOffer updated
