+++ src/main/java/cat/udl/eps/softarch/domain/BuyerCounterOffer.java
@@ -3,7 +3,11 @@ package cat.udl.eps.softarch.domain;
 /**
  * Created by xavier on 26/09/16.
  */
-public class BuyerCounterOffer extends Offer {
+public class BuyerCounterOffer extends Picture.Offer {
+
+    private String  agent;
+    private SellerOffer respondsTo;
+
     @Override
     public void setValue(float value){
         if (value < this.getValue()){
@@ -12,5 +16,21 @@ public class BuyerCounterOffer extends Offer {
         else{
             super.setValue(value);
         }
+    }
+
+    public String getAgent(){
+        return agent;
+    }
+
+    public void setAgent(String agent){
+        this.agent = agent;
+    }
+
+    public SellerOffer getRespondsTo(){
+        return respondsTo;
+    }
+
+    public void setRespondsTo(SellerOffer respondsTo){
+        this.respondsTo = respondsTo;
     }
 }
