+++ src/test/resources/features/Buyer Counter-Offer.feature
@@ -5,5 +5,5 @@ Feature: Buyer Counter-Offer
 
   Scenario: Create advertisement
     When I post a new buyer counter offer with a lower price value 100.0
-    Then A counter-offer id has been generated
+    Then A buyer counter-offer id has been generated
     And There is a new buyer counter offer with a lower price value 100.0
